pyproject.toml
[tool.poetry]
name = "livraria"
version = "0.1.0"
description = ""
authors = ["Brayan Luiz"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.8"
gunicorn = "^20.1.0"
whitenoise = "^5.3.0"
djangorestframework = "^3.12.4"
djangorestframework-simplejwt = "^5.0.0"
drf-spectacular = "^0.21.0"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
black = "^21.9b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
